+++ server/src/database/database.controller.ts
@@ -1,13 +1,4 @@
-import {
-  Body,
-  Controller,
-  Get,
-  Param,
-  Post,
-  Query,
-  Request,
-  UseInterceptors,
-} from '@nestjs/common';
+import { Controller } from '@nestjs/common';
 import { DatabaseService } from './database.service';
 
 // @UseInterceptors(SessionSerializer) // I think? (uncomment me when i work)
