+++ server/src/friend/friend.controller.ts
@@ -7,29 +7,37 @@ import {
   Post,
   Query,
   Request,
+  UseGuards,
   UseInterceptors,
 } from '@nestjs/common';
 import { Timetable, Settings, UserInterface } from 'src/schemas/user.schema';
 import { SessionSerializer } from 'src/auth/session.serializer';
-import { DatabaseService } from '../database/database.service';
+import { LoginGuard } from 'src/auth/login.guard';
 import { UserSettingsDto, UserTimetablesDto } from '../user/dtos/user.dto';
 
 import { FriendRequestDto } from './dtos/friend.dto';
 
 import { User } from '@sentry/node';
+import { FriendService } from './friend.service';
+import { UserService } from 'src/user/user.service';
 
 @Controller('friend')
 export class FriendController {
-  constructor(private readonly databaseService: DatabaseService) {}
+  constructor(
+    private readonly friendService: FriendService,
+    private readonly userService: UserService,
+  ) {}
 
+  @UseGuards(LoginGuard)
   @Get('/:userId')
   async getFriends(@Request() req): Promise<User | null> {
     if (req.query.userId) {
       // user id stuff here
-      return this.databaseService.getUser(req.query.userId);
+      return this.userService.getUser(req.query.userId);
     }
   }
 
+  @UseGuards(LoginGuard)
   @Post('/')
   async addFriend(
     @Request() req,
@@ -38,6 +46,7 @@ export class FriendController {
     return;
   }
 
+  @UseGuards(LoginGuard)
   @Delete('/')
   async deleteFriend(
     @Request() req,
@@ -46,6 +55,7 @@ export class FriendController {
     return;
   }
 
+  @UseGuards(LoginGuard)
   @Post('/request')
   async sendFriendRequest(
     @Request() req,
@@ -54,6 +64,7 @@ export class FriendController {
     return;
   }
 
+  @UseGuards(LoginGuard)
   @Post('/request')
   async deleteRequest(
     @Request() req,
